Edit Day control
HouseCollisonCheck: routine 트루일 때 frontDoor 활성화 안되게 / bed 조건 routine 추가 / frontDoor day에 따라 조건 설정

day 설정 : Day_manager(House 씬), / Day1_ctrl(프롤로그 끝난 후 게임 시작 할 때 day1)

out 설정: Out_ctrl(House 씬 내에서 0으로), HousecCollisionCheck(frontDoor 나가면 1로), Day1_ctrl( 프롤로그 끝난 후 게임 시작할 때 1로)

routine 설정: Day_manger(하루 시작할 때 0) / FadeOut_Day2End(Day2 끝나고 1로) / FadeOut_Day3_End(Day3 끝나고 1로)

+++ PBL_01/Assets/Day_manager.cs
@@ -4,6 +4,7 @@ using UnityEngine;
 
 public class Day_manager : MonoBehaviour
 {
+    int cnt = 0;
     // Start is called before the first frame update
     void Start()
     {
@@ -11,28 +12,33 @@ public class Day_manager : MonoBehaviour
         if( !PlayerPrefs.HasKey("bap") && !PlayerPrefs.HasKey("pill") && !PlayerPrefs.HasKey("planter")
             && !PlayerPrefs.HasKey("random1") && !PlayerPrefs.HasKey("random2") && !PlayerPrefs.HasKey("sleep"))
         {
-            PlayerPrefs.SetInt("bap", 0);
-            PlayerPrefs.SetInt("pill", 0);
-            PlayerPrefs.SetInt("planter", 0);
-            PlayerPrefs.SetInt("random1", 0);
-            PlayerPrefs.SetInt("random2", 0);
-            PlayerPrefs.SetInt("sleep", 0);
+            PlayerPrefs.SetInt("bap", 0); //밥먹기
+            PlayerPrefs.SetInt("pill", 0); //약먹기
+            PlayerPrefs.SetInt("planter", 0); //화분 가꾸기
+            PlayerPrefs.SetInt("random1", 0); //랜덤1
+            PlayerPrefs.SetInt("random2", 0); //랜덤2
+            PlayerPrefs.SetInt("routine", 0); //밖에서 일어나는 일
+            PlayerPrefs.SetInt("sleep", 0); //잠자기
         }
     }
 
     // Update is called once per frame
     void Update()
     {
-        //모두 진행되었을 때 초기화
+        //모두 진행되었을 때 키 삭제 초기화
         if ((bool)GetBool("bap") && (bool)GetBool("pill") && (bool)GetBool("planter")
-            && (bool)GetBool("random1") && (bool)GetBool("random2") && (bool)GetBool("sleep"))
+            && (bool)GetBool("random1") && (bool)GetBool("random2") && (bool)GetBool("routine") && (bool)GetBool("sleep"))
         {
             PlayerPrefs.DeleteKey("bap");
             PlayerPrefs.DeleteKey("pill");
             PlayerPrefs.DeleteKey("planter");
             PlayerPrefs.DeleteKey("random1");
             PlayerPrefs.DeleteKey("random2");
+            PlayerPrefs.DeleteKey("routine");
             PlayerPrefs.DeleteKey("sleep");
+
+            cnt += PlayerPrefs.GetInt("day") + 1;
+            PlayerPrefs.SetInt("day", cnt);
         }
     }
     // int 값이 1이면 true, 2이면 false
